Refactor : file directoring
