+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "korean-music-awards",
   "description": "",
-  "version": "2.2.3",
+  "version": "2.2.6",
   "type": "module",
   "license": "ISC",
   "main": "./dist/cjs/index.js",
