+++ src/utils/types/index.ts
@@ -1,0 +1,15 @@
+export interface MMADataType {
+  category: string;
+  artist: string;
+  title: string;
+  img: string;
+}
+
+export interface KHADataType {
+  category: string;
+  artist: string;
+  title: string;
+  img: string;
+  netizenRecommendations: string;
+  recommendationsList: string[];
+}
